fix: rename index.js component files
