+++ src/components/player/index.js
