+++ src/components/videocard/index.js
