+++ src/index.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.min.css'
-import Header from './components/header/header'
-import Dropzone from './components/dropzone/dropzone'
-import Sidebar from './components/sidebar/sidebar'
-import Player from './components/player/player'
-import Modal from './components/modal/modal'
+import Header from './components/header'
+import Dropzone from './components/dropzone'
+import Sidebar from './components/sidebar'
+import Player from './components/player'
+import Modal from './components/modal'
 import Playlist from './components/playlist/playlist'
 import { ToastContainer, toast, Bounce } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.min.css'
